ecr-upload/requirements.txt
matplotlib==3.7.1
numpy==1.25.0
Pillow==10.1.0
scikit_learn==1.3.2
torch==2.0.1
torchvision==0.15.0
opencv-python==4.8.1.78
scikit_image==0.22.0
boto3==1.29.2
botocore==1.32.2
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.18
requests